+++ TotL.Labyrinth/DeadEndCell.cs
@@ -1,26 +1,25 @@
-﻿using PathFinder;
-using PathFinder._2D;
-using SharpDX;
-using SharpDX.Toolkit.Graphics;
+﻿using PathFinder._2D;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using PathFinder;
+using SharpDX;
+using SharpDX.Toolkit.Graphics;
 
-namespace TotL.labyrinthcells
+namespace TotL.Labyrinth
 {
-    class TwoSideBlocked : Cell
+    public class DeadEndCell:Cell
     {
-        public TwoSideBlocked(int y,int x):base(y,x)
+        public DeadEndCell(int y,int x):base(y,x)
         {
-            texture = TextureFromFile.TextureProcessor.getTexture("twoSideBlockedCell");
+            texture = TextureFromFile.TextureProcessor.getTexture("DeadEndCell");
             setRotation(0f);
-            closedsides = 2;
+            closedsides = 3;
             X = x;
             Y = y;
         }
-
         public override void setRotation(float rotation)
         {
             base.rotation = rotation;
@@ -29,87 +28,87 @@ namespace TotL.labyrinthcells
                 up = true;
                 left = false;
                 down = false;
-                right = true;
+                right = false;
             }
             else
             if (rotation == Rotaitions.minus90)
             {
                 float unitSize = (Vars.ScreenWidth * 0.83f) / 25f;
                 LocationXoffset = 0;// unitSize;
-                up = true;
+                up = false;
                 left = true;
                 down = false;
                 right = false;
 
-            }else
+            }
+            else
             if (rotation == Rotaitions.half)
             {
                 float unitSize = (Vars.ScreenWidth * 0.83f) / 25f;
                 LocationXoffset = unitSize;
                 LocationYoffset = unitSize;
                 up = false;
-                left = true;
+                left = false;
                 down = true;
                 right = false;
 
-            }else
+            }
+            else
             if (rotation == Rotaitions.plus90)
             {
                 float unitSize = (Vars.ScreenWidth * 0.83f) / 25f;
                 LocationXoffset = unitSize;
                 up = false;
                 left = false;
-                down = true;
+                down = false;
                 right = true;
 
             }
-            
+
         }
-       
-        public override void update()
-        {
-           
-        }
+
         public override void draw()
         {
 
             Vars.spriteBatch.Draw(texture, new RectangleF(locationX, locationY, unitSize, unitSize), null, Color.White, rotation, new Vector2(0, 0), SpriteEffects.None, 0f);
 
 
-            /* foreach (var item in _blockedvolumes)
-             {
-                 Vars.spriteBatch.Draw(TextureFromFile.TextureProcessor.getTexture("transparent"), item, Color.White);
-             }*/
+            /*foreach (var item in _blockedvolumes)
+            {
+                Vars.spriteBatch.Draw(TextureFromFile.TextureProcessor.getTexture("transparent"), item, Color.White);
+            }*/
 
         }
 
         public override void SetBlockingVolumes()
         {
-            if (rotation==Rotaitions.zero)
+            if (rotation == Rotaitions.zero)
             {
-                _blockedvolumes.Add(new RectangleF(locationX,locationY,unitSize/4,unitSize));
-                _blockedvolumes.Add(new RectangleF(locationX+unitSize/4,(locationY+unitSize)-unitSize/4,unitSize-unitSize/4,unitSize/4));
-                _blockedvolumes.Add(new RectangleF((locationX+unitSize)-unitSize/4,locationY,unitSize/4,unitSize/4));
-            }else
-            if (rotation==Rotaitions.plus90)
+                _blockedvolumes.Add(new RectangleF(locationX, locationY, unitSize / 4, unitSize));
+                _blockedvolumes.Add(new RectangleF((locationX + unitSize) - unitSize / 4, locationY, unitSize / 4, unitSize));
+                _blockedvolumes.Add(new RectangleF(locationX + unitSize / 4, (locationY + unitSize) - unitSize / 4, unitSize - (2 * unitSize / 4), unitSize / 4));
+            }
+            else if (rotation == Rotaitions.plus90)
             {
-                _blockedvolumes.Add(new RectangleF(locationX-unitSize,locationY,unitSize,unitSize/4));
-                _blockedvolumes.Add(new RectangleF(locationX-unitSize,locationY+unitSize/4,unitSize/4,unitSize-unitSize/4));
-                _blockedvolumes.Add(new RectangleF(locationX-unitSize/4,(locationY+unitSize)-unitSize/4,unitSize/4,unitSize/4));
+                _blockedvolumes.Add(new RectangleF(locationX - unitSize, (locationY + unitSize) - unitSize / 4, unitSize, unitSize / 4));
+                _blockedvolumes.Add(new RectangleF(locationX - unitSize, locationY, unitSize, unitSize / 4));
+                _blockedvolumes.Add(new RectangleF(locationX - unitSize, locationY + unitSize / 4, unitSize / 4, unitSize - (2 * unitSize / 4)));
+            }
+            else if (rotation == Rotaitions.half)
+            {
+                _blockedvolumes.Add(new RectangleF((locationX - unitSize) + unitSize / 4, locationY - unitSize, unitSize - (2 * unitSize / 4), unitSize / 4));
+                _blockedvolumes.Add(new RectangleF(locationX - unitSize, locationY - unitSize, unitSize / 4, unitSize));
+                _blockedvolumes.Add(new RectangleF(locationX - unitSize / 4, locationY - unitSize, unitSize / 4, unitSize));
             }
             else if (rotation==Rotaitions.minus90)
             {
-                _blockedvolumes.Add(new RectangleF(locationX,locationY-unitSize,unitSize/4,unitSize/4));
-                _blockedvolumes.Add(new RectangleF((locationX+unitSize)-unitSize/4, locationY - unitSize, unitSize / 4,unitSize));
-                _blockedvolumes.Add(new RectangleF(locationX,locationY-unitSize/4,unitSize-unitSize/4,unitSize/4));
+                _blockedvolumes.Add(new RectangleF(locationX , locationY  - unitSize / 4, unitSize, unitSize / 4));
+                _blockedvolumes.Add(new RectangleF(locationX, locationY-unitSize, unitSize, unitSize / 4));
+                _blockedvolumes.Add(new RectangleF((locationX+unitSize)-unitSize/4, (locationY-unitSize) + unitSize / 4, unitSize / 4, unitSize - (2 * unitSize / 4)));
             }
-            else if (rotation==Rotaitions.half)
-            {
-                _blockedvolumes.Add(new RectangleF(locationX-unitSize,locationY-unitSize,unitSize,unitSize/4));
-                _blockedvolumes.Add(new RectangleF(locationX-unitSize,locationY-unitSize/4,unitSize/4,unitSize/4));
-                _blockedvolumes.Add(new RectangleF(locationX-unitSize/4,(locationY-unitSize)+unitSize/4,unitSize/4,unitSize-unitSize/4));
-            }
+
         }
+
 
         public override bool CheckFitting(Connection[,] connect, int co, int cs, int o, int s)
         {
@@ -181,6 +180,4 @@ namespace TotL.labyrinthcells
 
         }
     }
-    
-    
 }
