+++ TotL.Labyrinth/TunnelCell.cs
@@ -1,25 +1,27 @@
-﻿using PathFinder._2D;
+﻿using PathFinder;
+using PathFinder._2D;
+using SharpDX;
+using SharpDX.Toolkit.Graphics;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using PathFinder;
-using SharpDX;
-using SharpDX.Toolkit.Graphics;
 
-namespace TotL.labyrinthcells
+namespace TotL.Labyrinth
 {
-    class DeadEndCell:Cell
+    public class TunnelCell:Cell
     {
-        public DeadEndCell(int y,int x):base(y,x)
+
+        public TunnelCell(int y,int x):base(y,x)
         {
-            texture = TextureFromFile.TextureProcessor.getTexture("DeadEndCell");
+            texture = TextureFromFile.TextureProcessor.getTexture("TunnelCell");
             setRotation(0f);
-            closedsides = 3;
+            closedsides = 2;
             X = x;
             Y = y;
         }
+
         public override void setRotation(float rotation)
         {
             base.rotation = rotation;
@@ -27,7 +29,7 @@ namespace TotL.labyrinthcells
             {
                 up = true;
                 left = false;
-                down = false;
+                down = true;
                 right = false;
             }
             else
@@ -38,7 +40,7 @@ namespace TotL.labyrinthcells
                 up = false;
                 left = true;
                 down = false;
-                right = false;
+                right = true;
 
             }
             else
@@ -47,7 +49,7 @@ namespace TotL.labyrinthcells
                 float unitSize = (Vars.ScreenWidth * 0.83f) / 25f;
                 LocationXoffset = unitSize;
                 LocationYoffset = unitSize;
-                up = false;
+                up = true;
                 left = false;
                 down = true;
                 right = false;
@@ -59,7 +61,7 @@ namespace TotL.labyrinthcells
                 float unitSize = (Vars.ScreenWidth * 0.83f) / 25f;
                 LocationXoffset = unitSize;
                 up = false;
-                left = false;
+                left = true;
                 down = false;
                 right = true;
 
@@ -72,42 +74,27 @@ namespace TotL.labyrinthcells
 
             Vars.spriteBatch.Draw(texture, new RectangleF(locationX, locationY, unitSize, unitSize), null, Color.White, rotation, new Vector2(0, 0), SpriteEffects.None, 0f);
 
-
-            /*foreach (var item in _blockedvolumes)
-            {
-                Vars.spriteBatch.Draw(TextureFromFile.TextureProcessor.getTexture("transparent"), item, Color.White);
-            }*/
+            /*  foreach (var item in _blockedvolumes)
+              {
+                  Vars.spriteBatch.Draw(TextureFromFile.TextureProcessor.getTexture("transparent"), item, Color.White);
+              }*/
 
         }
 
         public override void SetBlockingVolumes()
         {
-            if (rotation == Rotaitions.zero)
+            if (rotation == Rotaitions.zero || rotation == Rotaitions.half)
             {
                 _blockedvolumes.Add(new RectangleF(locationX, locationY, unitSize / 4, unitSize));
                 _blockedvolumes.Add(new RectangleF((locationX + unitSize) - unitSize / 4, locationY, unitSize / 4, unitSize));
-                _blockedvolumes.Add(new RectangleF(locationX + unitSize / 4, (locationY + unitSize) - unitSize / 4, unitSize - (2 * unitSize / 4), unitSize / 4));
             }
-            else if (rotation == Rotaitions.plus90)
+            else if (rotation == Rotaitions.minus90 || rotation == Rotaitions.plus90)
             {
-                _blockedvolumes.Add(new RectangleF(locationX - unitSize, (locationY + unitSize) - unitSize / 4, unitSize, unitSize / 4));
-                _blockedvolumes.Add(new RectangleF(locationX - unitSize, locationY, unitSize, unitSize / 4));
-                _blockedvolumes.Add(new RectangleF(locationX - unitSize, locationY + unitSize / 4, unitSize / 4, unitSize - (2 * unitSize / 4)));
+                _blockedvolumes.Add(new RectangleF(locationX-unitSize, locationY, unitSize, unitSize / 4));
+                _blockedvolumes.Add(new RectangleF(locationX-unitSize,(locationY+unitSize)-unitSize/4,unitSize,unitSize/4));
             }
-            else if (rotation == Rotaitions.half)
-            {
-                _blockedvolumes.Add(new RectangleF((locationX - unitSize) + unitSize / 4, locationY - unitSize, unitSize - (2 * unitSize / 4), unitSize / 4));
-                _blockedvolumes.Add(new RectangleF(locationX - unitSize, locationY - unitSize, unitSize / 4, unitSize));
-                _blockedvolumes.Add(new RectangleF(locationX - unitSize / 4, locationY - unitSize, unitSize / 4, unitSize));
-            }
-            else if (rotation==Rotaitions.minus90)
-            {
-                _blockedvolumes.Add(new RectangleF(locationX , locationY  - unitSize / 4, unitSize, unitSize / 4));
-                _blockedvolumes.Add(new RectangleF(locationX, locationY-unitSize, unitSize, unitSize / 4));
-                _blockedvolumes.Add(new RectangleF((locationX+unitSize)-unitSize/4, (locationY-unitSize) + unitSize / 4, unitSize / 4, unitSize - (2 * unitSize / 4)));
-            }
-
         }
+            
 
 
         public override bool CheckFitting(Connection[,] connect, int co, int cs, int o, int s)
