+++ TotL.Labyrinth/TwoSideBlocked.cs
@@ -1,55 +1,114 @@
-﻿using System;
+﻿using PathFinder;
+using PathFinder._2D;
+using SharpDX;
+using SharpDX.Toolkit.Graphics;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using PathFinder._2D;
-using SharpDX.Toolkit;
-using SharpDX.Toolkit.Graphics;
-using PathFinder;
-using SharpDX;
 
-namespace TotL.labyrinthcells
+namespace TotL.Labyrinth
 {
-    class CrossCell : Cell 
+    public class TwoSideBlocked : Cell
     {
-        public CrossCell(int y,int x):base(y,x)
+        public TwoSideBlocked(int y,int x):base(y,x)
         {
-            texture = TextureFromFile.TextureProcessor.getTexture("CrossCell");
-            closedsides = 0;
+            texture = TextureFromFile.TextureProcessor.getTexture("twoSideBlockedCell");
+            setRotation(0f);
+            closedsides = 2;
             X = x;
             Y = y;
         }
 
-
         public override void setRotation(float rotation)
         {
-            base.rotation = 0;
-           
-            up = true;
-            left = true;
-            down = true;
-            right = true;
+            base.rotation = rotation;
+            if (rotation == Rotaitions.zero)
+            {
+                up = true;
+                left = false;
+                down = false;
+                right = true;
+            }
+            else
+            if (rotation == Rotaitions.minus90)
+            {
+                float unitSize = (Vars.ScreenWidth * 0.83f) / 25f;
+                LocationXoffset = 0;// unitSize;
+                up = true;
+                left = true;
+                down = false;
+                right = false;
+
+            }else
+            if (rotation == Rotaitions.half)
+            {
+                float unitSize = (Vars.ScreenWidth * 0.83f) / 25f;
+                LocationXoffset = unitSize;
+                LocationYoffset = unitSize;
+                up = false;
+                left = true;
+                down = true;
+                right = false;
+
+            }else
+            if (rotation == Rotaitions.plus90)
+            {
+                float unitSize = (Vars.ScreenWidth * 0.83f) / 25f;
+                LocationXoffset = unitSize;
+                up = false;
+                left = false;
+                down = true;
+                right = true;
+
+            }
             
         }
-
+       
+        public override void update()
+        {
+           
+        }
         public override void draw()
         {
-            
+
             Vars.spriteBatch.Draw(texture, new RectangleF(locationX, locationY, unitSize, unitSize), null, Color.White, rotation, new Vector2(0, 0), SpriteEffects.None, 0f);
-           /* foreach (var item in _blockedvolumes)
-            {
-                Vars.spriteBatch.Draw(TextureFromFile.TextureProcessor.getTexture("transparent"), item, Color.White);
-            }*/
+
+
+            /* foreach (var item in _blockedvolumes)
+             {
+                 Vars.spriteBatch.Draw(TextureFromFile.TextureProcessor.getTexture("transparent"), item, Color.White);
+             }*/
 
         }
+
         public override void SetBlockingVolumes()
         {
-           
-            _blockedvolumes.Add(new RectangleF((locationX + unitSize) - unitSize / 4, locationY, unitSize / 4, unitSize / 4));
-            _blockedvolumes.Add(new RectangleF((locationX + unitSize) - unitSize / 4, (locationY + unitSize) - unitSize / 4, unitSize / 4, unitSize / 4));
-            _blockedvolumes.Add(new RectangleF(locationX, locationY, unitSize / 4, unitSize / 4));
-            _blockedvolumes.Add(new RectangleF(locationX , (locationY + unitSize) - unitSize / 4, unitSize / 4, unitSize / 4));
+            if (rotation==Rotaitions.zero)
+            {
+                _blockedvolumes.Add(new RectangleF(locationX,locationY,unitSize/4,unitSize));
+                _blockedvolumes.Add(new RectangleF(locationX+unitSize/4,(locationY+unitSize)-unitSize/4,unitSize-unitSize/4,unitSize/4));
+                _blockedvolumes.Add(new RectangleF((locationX+unitSize)-unitSize/4,locationY,unitSize/4,unitSize/4));
+            }else
+            if (rotation==Rotaitions.plus90)
+            {
+                _blockedvolumes.Add(new RectangleF(locationX-unitSize,locationY,unitSize,unitSize/4));
+                _blockedvolumes.Add(new RectangleF(locationX-unitSize,locationY+unitSize/4,unitSize/4,unitSize-unitSize/4));
+                _blockedvolumes.Add(new RectangleF(locationX-unitSize/4,(locationY+unitSize)-unitSize/4,unitSize/4,unitSize/4));
+            }
+            else if (rotation==Rotaitions.minus90)
+            {
+                _blockedvolumes.Add(new RectangleF(locationX,locationY-unitSize,unitSize/4,unitSize/4));
+                _blockedvolumes.Add(new RectangleF((locationX+unitSize)-unitSize/4, locationY - unitSize, unitSize / 4,unitSize));
+                _blockedvolumes.Add(new RectangleF(locationX,locationY-unitSize/4,unitSize-unitSize/4,unitSize/4));
+            }
+            else if (rotation==Rotaitions.half)
+            {
+                _blockedvolumes.Add(new RectangleF(locationX-unitSize,locationY-unitSize,unitSize,unitSize/4));
+                _blockedvolumes.Add(new RectangleF(locationX-unitSize,locationY-unitSize/4,unitSize/4,unitSize/4));
+                _blockedvolumes.Add(new RectangleF(locationX-unitSize/4,(locationY-unitSize)+unitSize/4,unitSize/4,unitSize-unitSize/4));
+            }
         }
 
         public override bool CheckFitting(Connection[,] connect, int co, int cs, int o, int s)
@@ -122,4 +181,6 @@ namespace TotL.labyrinthcells
 
         }
     }
+    
+    
 }
