+++ TotL.Labyrinth/packages.config
@@ -1,0 +1,10 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="SharpDX" version="2.6.3" targetFramework="net452" />
+  <package id="SharpDX.Direct3D11" version="2.6.3" targetFramework="net452" />
+  <package id="SharpDX.DXGI" version="2.6.3" targetFramework="net452" />
+  <package id="SharpDX.Toolkit" version="2.6.3" targetFramework="net452" />
+  <package id="SharpDX.Toolkit.Game" version="2.6.3" targetFramework="net452" />
+  <package id="SharpDX.Toolkit.Graphics" version="2.6.3" targetFramework="net452" />
+  <package id="SharpDX.Toolkit.Input" version="2.6.3" targetFramework="net452" />
+</packages>
