+++ TotL.sln
@@ -9,6 +9,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PathFinder", "PathFinder\Pa
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PathFinder.AStar", "PathFinder.Astar\PathFinder.AStar\PathFinder.AStar.csproj", "{F1AF3392-22D1-4B03-9C3B-C10DA9B1BCCD}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TotL.Labyrinth", "TotL.Labyrinth\TotL.Labyrinth.csproj", "{2E89AC9D-D468-4D35-A7FA-29C8FAC33956}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -40,6 +42,14 @@ Global
 		{F1AF3392-22D1-4B03-9C3B-C10DA9B1BCCD}.Release|Any CPU.ActiveCfg = Release|x86
 		{F1AF3392-22D1-4B03-9C3B-C10DA9B1BCCD}.Release|x86.ActiveCfg = Release|x86
 		{F1AF3392-22D1-4B03-9C3B-C10DA9B1BCCD}.Release|x86.Build.0 = Release|x86
+		{2E89AC9D-D468-4D35-A7FA-29C8FAC33956}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2E89AC9D-D468-4D35-A7FA-29C8FAC33956}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2E89AC9D-D468-4D35-A7FA-29C8FAC33956}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{2E89AC9D-D468-4D35-A7FA-29C8FAC33956}.Debug|x86.Build.0 = Debug|Any CPU
+		{2E89AC9D-D468-4D35-A7FA-29C8FAC33956}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2E89AC9D-D468-4D35-A7FA-29C8FAC33956}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2E89AC9D-D468-4D35-A7FA-29C8FAC33956}.Release|x86.ActiveCfg = Release|Any CPU
+		{2E89AC9D-D468-4D35-A7FA-29C8FAC33956}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
